Merch_Microservices_App/backend/auth/requirements.txt
fastapi
uvicorn
mongoengine
pydantic
python-multipart==0.0.5